requirements.txt
fastapi
uvicorn
python-jose
beautifulsoup4